refactor(overlay): 优化Overlay组件目录结构
