+++ packages/devui-vue/devui/overlay/index.ts
@@ -2,7 +2,9 @@ import type { App } from 'vue';
 import { FixedOverlay } from './src/fixed-overlay';
 import { FlexibleOverlay } from './src/flexible-overlay';
 import { inBrowser } from '../shared/util/common-var';
-export * from './src/overlay-types';
+
+export * from './src/fixed-overlay/fixed-overlay-types';
+export * from './src/flexible-overlay/flexible-overlay-types';
 
 export { FlexibleOverlay, FixedOverlay };
