+++ packages/devui-vue/devui/overlay/src/fixed-overlay/fixed-overlay.scss
@@ -1,0 +1,22 @@
+.devui-overlay-background {
+  position: fixed;
+  top: 0;
+  left: 0;
+  height: 100vh;
+  width: 100vw;
+  display: flex;
+
+  &__color {
+    background: rgba(0, 0, 0, 0.4);
+  }
+
+  .devui-overlay {
+    position: relative;
+    z-index: 1000;
+    pointer-events: auto;
+  }
+
+  &__disabled {
+    pointer-events: none;
+  }
+}
