+++ packages/devui-vue/devui/overlay/src/flexible-overlay/flexible-overlay.scss
@@ -1,4 +1,4 @@
-@import '../../styles-var/devui-var.scss';
+@import '../../../styles-var/devui-var.scss';
 
 .devui-flexible-overlay {
   position: fixed;
